llms/agents/BQ_test_1/pyproject.toml
[project]
name = "bq-nl2sql-agent"
version = "0.1.0"
description = "BigQuery NL2SQL agent built with Google ADK — translates natural language to SQL."
requires-python = ">=3.11"

dependencies = [
    # Google ADK core
    "google-adk>=1.0",
    # BigQuery client (schema discovery)
    "google-cloud-bigquery>=3.0",
    # BigQuery type support (db-dtypes needed for ADK BigQueryToolset)
    "db-dtypes>=1.4",
    # Vertex AI SDK — required for Agent Engine deployment
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
    # Env-var loading
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.26",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bq_agent"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
